pyproject.toml
[tool.poetry]
name = "globalclasssearcher"
version = "0.1.0"
description = ""
authors = ["mmonj <mmonj@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.12.1"
Django = "5.0.2"
reactivated = "0.44.4"
psycopg2-binary = "^2.9.9"
django-stubs = "^5.0.2"
django-stubs-ext = "^5.0.2"
django-tasks-scheduler = "^2.1.0"
requests = "^2.32.3"
djangorestframework = "^3.15.2"
beautifulsoup4 = "^4.12.3"
lxml = "^5.3.0"
cattrs = "^25.1.0"
natsort = "^8.4.0"


[tool.poetry.group.dev.dependencies]
djangorestframework-stubs = "^3.15.1"
types-beautifulsoup4 = "^4.12.0.20241020"
ruff = "^0.7.3"
types-pytz = "^2025.1.0.20250204"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
